requirements.txt
altair==4.1.0
pandas==1.3.2
psycopg2-binary==2.9.1
streamlit==1.0.0
streamlit-aggrid==0.2.2.post2
